Spring security

+++ src/main/java/ar/edu/unq/grupof/desarrollo_app_criptop2p/service/TransactionService.java
@@ -8,5 +8,5 @@ import java.util.Optional;
 public interface TransactionService {
     List<Transaction> findAllTransactions();
     Transaction saveTransaction(Transaction transaction);
-    Transaction getTransactionById(Integer id);
+    Transaction getTransactionById(Long id);
 }
